refactor: Remove unused code and update leaderboard generation
- Remove unused code in the `route.ts` file
- Update the `leaderboard.ts` file to generate the leaderboard using data from the database instead of an external API
- Add a new file `leaderboard.tsx` to generate the leaderboard card image
